lambda/user/common/python/requirements.txt
mysql_connector_python